pyproject.toml
[project]
name = "llm-finetuning"
version = "0.1.0"
description = "LLM Supervised Fine-Tuning with LoRA/QLoRA optimized for Apple Silicon"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "torch>=2.1.0",
    "transformers>=4.36.0",
    "accelerate>=0.25.0",
    "peft>=0.7.0",
    "datasets>=2.15.0",
    "trl>=0.7.4",
    "scipy>=1.11.4",
    "evaluate>=0.4.1",
    "scikit-learn>=1.3.2",
    "pandas>=2.1.4",
    "numpy>=1.26.2",
    "wandb>=0.16.1",
    "tensorboard>=2.15.1",
    "tqdm>=4.66.1",
    "pyyaml>=6.0.1",
    "sentencepiece>=0.1.99",
    "protobuf>=4.25.1",
    "matplotlib>=3.9.4",
    "seaborn>=0.13.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.hatch.build.targets.wheel]
packages = ["."]

